+++ src/main/java/ru/mbt_bs/fluffy_crm/services/CustomerService.java
@@ -6,7 +6,7 @@ import ru.mbt_bs.fluffy_crm.data.json.CustomerLink;
 import java.util.List;
 
 public interface CustomerService {
-    Customer getCustomer(Integer id);
+    Customer getCustomer(Long id);
     void updateCustomer(Customer customer);
     List<CustomerLink> getCustomers(String name);
 }
